Align Jedis target helper responsibilities
