+++ instrumentation/jedis/jedis-4.0/javaagent-unit-tests/src/test/java/io/opentelemetry/javaagent/instrumentation/jedis/v4_0/JedisServerTargetTest.java
@@ -13,14 +13,14 @@ import java.util.List;
 import org.junit.jupiter.api.Test;
 import redis.clients.jedis.HostAndPort;
 
-class JedisConfiguredTargetsTest {
+class JedisServerTargetTest {
 
   private static final HostAndPort NODE_ONE = new HostAndPort("node1", 6379);
   private static final HostAndPort NODE_TWO = new HostAndPort("node2", 6380);
 
   @Test
   void clusterNodesAreSorted() {
-    RedisServerTarget target = JedisSingletons.targetOfNodes(asList(NODE_TWO, NODE_ONE));
+    RedisServerTarget target = JedisServerTarget.ofNodes(asList(NODE_TWO, NODE_ONE));
 
     assertThat(target).isNotNull();
     assertThat(target.getAddress()).isEqualTo("node1:6379,node2:6380");
@@ -29,7 +29,7 @@ class JedisConfiguredTargetsTest {
 
   @Test
   void shardsPreserveConfiguredOrder() {
-    RedisServerTarget target = JedisSingletons.targetOfShards(asList(NODE_TWO, NODE_ONE));
+    RedisServerTarget target = JedisServerTarget.ofShards(asList(NODE_TWO, NODE_ONE));
 
     assertThat(target).isNotNull();
     assertThat(target.getAddress()).isEqualTo("node2:6380,node1:6379");
@@ -39,7 +39,7 @@ class JedisConfiguredTargetsTest {
   @Test
   void sentinelsAreSanitizedAndIncludeTheMasterName() {
     RedisServerTarget target =
-        JedisSingletons.targetOfSentinels(
+        JedisServerTarget.ofSentinels(
             "mymaster",
             asList(new HostAndPort("node2", 26379), "redis://user:secret@node1:26379/0"));
 
@@ -52,7 +52,7 @@ class JedisConfiguredTargetsTest {
   void unrepresentableMemberDropsTheWholeTarget() {
     List<Object> nodes = asList(NODE_ONE, new Object());
 
-    assertThat(JedisSingletons.targetOfNodes(nodes)).isNull();
-    assertThat(JedisSingletons.targetOfSentinels("mymaster", nodes)).isNull();
+    assertThat(JedisServerTarget.ofNodes(nodes)).isNull();
+    assertThat(JedisServerTarget.ofSentinels("mymaster", nodes)).isNull();
   }
 }
