+++ instrumentation/jedis/jedis-4.0/javaagent/src/main/java/io/opentelemetry/javaagent/instrumentation/jedis/v4_0/JedisConfiguredTargets.java
@@ -70,7 +70,7 @@ public class JedisConfiguredTargets {
 
   public static void setTopologyTargetFromNodes(
       JedisClusterInfoCache topologyOwner, Collection<?> startNodes) {
-    setTopologyTarget(topologyOwner, JedisSingletons.targetOfNodes(startNodes));
+    setTopologyTarget(topologyOwner, JedisServerTarget.ofNodes(startNodes));
   }
 
   @Nullable
