+++ instrumentation/jedis/jedis-4.0/javaagent/src/main/java/io/opentelemetry/javaagent/instrumentation/jedis/v4_0/JedisSingletons.java
@@ -11,17 +11,12 @@ import io.opentelemetry.api.GlobalOpenTelemetry;
 import io.opentelemetry.instrumentation.api.incubator.semconv.db.DbClientAttributesExtractor;
 import io.opentelemetry.instrumentation.api.incubator.semconv.db.DbClientMetrics;
 import io.opentelemetry.instrumentation.api.incubator.semconv.db.DbClientSpanNameExtractor;
-import io.opentelemetry.instrumentation.api.incubator.semconv.db.internal.RedisServerTarget;
 import io.opentelemetry.instrumentation.api.instrumenter.Instrumenter;
 import io.opentelemetry.instrumentation.api.instrumenter.InstrumenterBuilder;
 import io.opentelemetry.instrumentation.api.instrumenter.SpanKindExtractor;
 import io.opentelemetry.instrumentation.api.util.VirtualField;
-import java.util.ArrayList;
-import java.util.Collection;
-import java.util.List;
 import javax.annotation.Nullable;
 import redis.clients.jedis.Connection;
-import redis.clients.jedis.HostAndPort;
 import redis.clients.jedis.JedisSocketFactory;
 
 public class JedisSingletons {
@@ -68,53 +63,6 @@ public class JedisSingletons {
   public static void setConnectionInfo(
       Connection connection, JedisSocketFactory socketFactory, @Nullable Object clientConfig) {
     CONNECTION_INFO.set(connection, JedisConnectionInfo.create(socketFactory, clientConfig));
-  }
-
-  @Nullable
-  public static RedisServerTarget targetOfNodes(@Nullable Collection<?> nodes) {
-    return RedisServerTarget.ofUnorderedEndpoints(endpointStrings(nodes));
-  }
-
-  @Nullable
-  public static RedisServerTarget targetOfShards(@Nullable List<HostAndPort> shards) {
-    return RedisServerTarget.ofEndpoints(endpointStrings(shards));
-  }
-
-  @Nullable
-  public static RedisServerTarget targetOfSentinels(
-      @Nullable String masterName, @Nullable Collection<?> sentinels) {
-    List<String> endpoints = null;
-    if (sentinels != null) {
-      endpoints = new ArrayList<>(sentinels.size());
-      for (Object sentinel : sentinels) {
-        if (sentinel instanceof HostAndPort) {
-          HostAndPort hostAndPort = (HostAndPort) sentinel;
-          endpoints.add(RedisServerTarget.endpoint(hostAndPort.getHost(), hostAndPort.getPort()));
-        } else if (sentinel instanceof String) {
-          endpoints.add(RedisServerTarget.normalizeHostAndPort((String) sentinel));
-        } else {
-          endpoints.add(null);
-        }
-      }
-    }
-    return RedisServerTarget.ofUnorderedEndpointsAndLogicalName(endpoints, masterName);
-  }
-
-  @Nullable
-  private static List<String> endpointStrings(@Nullable Collection<?> nodes) {
-    if (nodes == null || nodes.isEmpty()) {
-      return null;
-    }
-    List<String> endpoints = new ArrayList<>(nodes.size());
-    for (Object value : nodes) {
-      if (value instanceof HostAndPort) {
-        HostAndPort node = (HostAndPort) value;
-        endpoints.add(RedisServerTarget.endpoint(node.getHost(), node.getPort()));
-      } else {
-        endpoints.add(null);
-      }
-    }
-    return endpoints;
   }
 
   private JedisSingletons() {}
